# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Paadi backend — auth, pots, payments, settlements",
    "title": "Paadi API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/developer/webhooks": {
      "get": {
        "description": "Returns the caller's registered webhook endpoints. Signing secrets are never included in this response.",
        "operationId": "DeveloperController_list",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "endpoints": {
                      "items": {
                        "properties": {
                          "createdAt": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "events": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "id": {
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "ACTIVE",
                              "DISABLED"
                            ],
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "url": {
                            "format": "uri",
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "url",
                          "events",
                          "description",
                          "status",
                          "createdAt",
                          "updatedAt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "endpoints"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List webhook endpoints",
        "tags": [
          "developer"
        ]
      },
      "post": {
        "description": "Registers a URL to receive signed event deliveries. The signing secret is generated server-side and returned once in this response.",
        "operationId": "DeveloperController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "description": {
                    "maxLength": 280,
                    "type": "string"
                  },
                  "events": {
                    "items": {
                      "enum": [
                        "wallet.credited",
                        "wallet.debited",
                        "withdrawal.completed",
                        "exception.resolved"
                      ],
                      "type": "string"
                    },
                    "minItems": 1,
                    "type": "array"
                  },
                  "url": {
                    "format": "uri",
                    "type": "string"
                  }
                },
                "required": [
                  "url",
                  "events"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "events": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "secret": {
                      "type": "string"
                    },
                    "status": {
                      "enum": [
                        "ACTIVE",
                        "DISABLED"
                      ],
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "url": {
                      "format": "uri",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "url",
                    "events",
                    "description",
                    "status",
                    "createdAt",
                    "updatedAt",
                    "secret"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register webhook endpoint",
        "tags": [
          "developer"
        ]
      }
    },
    "/developer/webhooks/{id}": {
      "delete": {
        "description": "Soft-disables a webhook endpoint the caller owns. No further deliveries are attempted to it.",
        "operationId": "DeveloperController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "createdAt": {
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "events": {
                      "items": {
                        "type": "string"
                      },
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "enum": [
                        "ACTIVE",
                        "DISABLED"
                      ],
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "url": {
                      "format": "uri",
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "url",
                    "events",
                    "description",
                    "status",
                    "createdAt",
                    "updatedAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      ]
                    },
                    "statusCode": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable webhook endpoint",
        "tags": [
          "developer"
        ]
      }
    },
...
  },
  "servers": [],
...
}